Use explicit template instantiation
